refactor: rename logger/runTrace.ts -> channelTrace.ts to remove naming collision
`src/logger/` and `src/transcript/` both contained a file named `runTrace.ts`,
even though they serve different purposes. The logger file creates bare
channel-only traces (no transcript); the transcript file creates the full
product-level run trace. The identical filenames made it easy to open the
wrong file and obscured the architectural boundary between the SDK-safe logger
layer and the product transcript layer. Rename to match the existing exported
function name (`createChannelTrace` -> `channelTrace.ts`); all callers already
use the `@logger` barrel so no import sites change.
